requirements.txt
pip-tools
ipython
ipdb
flake8
Django
django-suit
django-reversion
django-suit-redactor
django-mptt
django-taggit
psycopg2
raven
WebTest
django-webtest
model-mommy
django-debug-toolbar
django-model-utils
django-redis
hiredis
django-bootstrap3
django-vanilla-views
django-extensions
django-clear-cache
uwsgi
django-configurations
tablib
pymongo
tika
tqdm
django-attachments
django-haystack
pysolr
arrow
python-dateutil
-e git+git://github.com/pashinin/fabric@p33#egg=fabric

